chore; move .get() back to feeder.go file
